refactor(utils): move findInField to its own file
